requirements.txt
streamlit
google-generativeai
scikit-learn
numpy
watchdog
youtube-transcript-api